// apps/web/src/shared/api/generated/model/leagueRankRowDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Gravit API Docs
 * 앱센터 16.5기 동계 프로젝트 Gravit API Docs
 * OpenAPI spec version: 1.0.0
 */

export interface LeagueRankRowDto {
  rank?: number;
  userId?: number;
  lp?: number;
  nickname: string;
  profileImgNumber?: number;
  xp?: number;
  level?: number;
}